+++ utils/configuration.py
@@ -1,6 +1,8 @@
 """Handles loading default and user configuration from YAML."""
 
 # Standard imports
+import urllib.request
+from pathlib import Path
 from typing import Any, Literal
 
 # External imports
@@ -12,21 +14,31 @@ CONFIG_PATHS: dict[str, str] = {
     'default': 'config_default.yml'
 }
 
+if not Path(CONFIG_PATHS['default']).is_file():
+    print('config_default.yml not found; downloading the latest version...')
+    urllib.request.urlretrieve('https://raw.githubusercontent.com/svioletg/viMusBot/master/config_default.yml',
+        Path(CONFIG_PATHS['default']))
+
+if not Path(CONFIG_PATHS['user']).is_file():
+    print('config.yml does not exist; creating blank config.yml...')
+    with open(CONFIG_PATHS['user'], 'w', encoding='utf-8') as f:
+        f.write('')
+
 with open(CONFIG_PATHS['default'], 'r', encoding='utf-8') as f:
-    config_default = benedict(yaml.safe_load(f))
+    CONFIG_DEFAULT = benedict(yaml.safe_load(f))
 
 with open(CONFIG_PATHS['user'], 'r', encoding='utf-8') as f:
-    config_yaml = yaml.safe_load(f)
-    config = benedict(config_yaml) if config_yaml else benedict({})
+    CONFIG_YAML = yaml.safe_load(f)
+    CONFIG = benedict(CONFIG_YAML) if CONFIG_YAML else benedict({})
 
 def get(key: str) -> Any:
     """Looks for the given key in the user config, returns the default value if none is set."""
-    return config.get(key, config_default.get(key))
+    return CONFIG.get(key, CONFIG_DEFAULT.get(key))
 
 def get_default(key: str) -> Any:
     """Returns the default value for a given key."""
-    return config_default.get(key)
+    return CONFIG_DEFAULT.get(key)
 
 def get_full(config_type: Literal['user', 'default']) -> benedict:
     """Returns the full config benedict object."""
-    return config if config_type == 'user' else config_default
+    return CONFIG if config_type == 'user' else CONFIG_DEFAULT